Update valueChanged for strict parsing with format
Was not working with 'DD/MM/YYYY hh:mm A' format. This change works well for me, but I have not tested it. But as I can see in line 285, it already takes care of that moment date to be valid. Then I think it will work for other use cases.