Fix order-insensitive array comparison (#75)
* Fix order-insensitive array comparison

When multiselect is true and "Select all" is clicked, there is an infinite loop where bootstrap-select returns a list, the list is set in abp-select that causes bootstrap-select to fire again, it returns the list to abp-select, but in a different order, etc. etc.  The commented out code that uses a.every and b.includes works in typescript.  I'm assuming you are using Babel, and I don't know if it will polyfill for old IE browsers, so I dropped back to plain old Javascript looping.

* Update util-service.js

* Update util-service.js

* lint, fix indentation